Pull downloader into its own file
